+++ src/components/Nav.tsx
@@ -1,31 +1,18 @@
 import { css } from '@emotion/css';
 import { GrafanaTheme2 } from '@grafana/data';
-import { Button } from '@grafana/ui';
 
-import { tryOrReport } from '../../helpers';
-import { fmtCue, fmtJson } from '../../services/format';
-import { useInputContext, useLineageContext } from '../../state';
-import { useStyles } from '../../theme';
-import Examples from './Examples';
-import OpSelector from './OpSelector';
-import Share from './Share';
-import ThemeSwitch from './ThemeSwitch';
+import { useStyles } from '../theme';
+import ThemeSwitch from './Actions/ThemeSwitch';
 
-const Header = () => {
+export const Nav = () => {
   const styles = useStyles(getStyles);
   return (
-    <div className={styles.header}>
-      <div className={styles.flex}>
+    <nav className={styles.nav}>
+      <div className={styles.section}>
         <img src={`${process.env.PUBLIC_URL}/grafana.svg`} alt={'Grafana logo'} className={styles.logo} />
         <h4 className={styles.headerText}>Thema Playground</h4>
       </div>
-      <div className={styles.flex}>
-        <div className={styles.flex}>
-          <Examples />
-          <Share />
-          <FormatButton />
-          <OpSelector />
-        </div>
+      <div className={styles.section}>
         <a
           className={styles.icon}
           href={'https://github.com/grafana/play-thema'}
@@ -45,39 +32,13 @@ const Header = () => {
         </a>
         <ThemeSwitch className={styles.themeSwitch} />
       </div>
-    </div>
-  );
-};
-
-const FormatButton = () => {
-  const { input, setInput } = useInputContext();
-  const { lineage, setLineage } = useLineageContext();
-
-  const formatFn = () => {
-    tryOrReport(() => {
-      setInput(fmtJson(input));
-      setLineage(fmtCue(lineage));
-    }, true);
-  };
-
-  return (
-    <Button onClick={formatFn} variant={'secondary'}>
-      Format
-    </Button>
+    </nav>
   );
 };
 
 const getStyles = (theme: GrafanaTheme2) => {
   return {
-    flex: css`
-      display: flex;
-      align-items: center;
-
-      & > * {
-        margin-right: ${theme.spacing(2)};
-      }
-    `,
-    header: css`
+    nav: css`
       display: flex;
       align-items: center;
       justify-content: space-between;
@@ -89,12 +50,16 @@ const getStyles = (theme: GrafanaTheme2) => {
       border-radius: 2px;
       color: ${theme.colors.text.primary};
     `,
+    section: css`
+      display: flex;
+      align-items: center;
+
+      & > * {
+        margin-right: ${theme.spacing(2)};
+      }
+    `,
     headerText: css`
       margin-bottom: 0;
-    `,
-    themeSwitch: css`
-      color: ${theme.colors.text.primary};
-      cursor: pointer;
     `,
     logo: css`
       width: 20px;
@@ -104,6 +69,9 @@ const getStyles = (theme: GrafanaTheme2) => {
     icon: css`
       color: ${theme.colors.text.primary};
     `,
+    themeSwitch: css`
+      color: ${theme.colors.text.primary};
+      cursor: pointer;
+    `,
   };
 };
-export default Header;
